Implement role based security and check if active user is authorized to execute method

+++ backmeup-service-rest/src/main/java/org/backmeup/rest/resources/Users.java
@@ -3,6 +3,8 @@ package org.backmeup.rest.resources;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.annotation.security.PermitAll;
+import javax.annotation.security.RolesAllowed;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
@@ -12,17 +14,26 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
+import javax.ws.rs.WebApplicationException;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response.Status;
+import javax.ws.rs.core.SecurityContext;
 
 import org.backmeup.model.BackMeUpUser;
 import org.backmeup.model.dto.UserDTO;
 import org.backmeup.rest.DummyDataManager;
+import org.backmeup.rest.auth.BackmeupPrincipal;
 
 /**
  * All user specific operations will be handled within this class.
  */
 @Path("/users")
 public class Users extends Base {	
+	@Context
+    private SecurityContext securityContext;
+	
+	@RolesAllowed("user")
 	@GET
 	@Path("/")
 	@Produces(MediaType.APPLICATION_JSON)
@@ -33,6 +44,7 @@ public class Users extends Base {
 		return userList;
 	}
 	
+	@PermitAll
 	@POST
 	@Path("/")
 	@Consumes(MediaType.APPLICATION_JSON)
@@ -41,36 +53,52 @@ public class Users extends Base {
 		BackMeUpUser userModel = getMapper().map(user, BackMeUpUser.class);
 		userModel = getLogic().addUser(userModel);
 		return getMapper().map(userModel, UserDTO.class);
-//		return DummyDataManager.getUserDTO();
 	}
 	
+	@RolesAllowed("user")
 	@GET
 	@Path("/{userId}")
 	@Produces(MediaType.APPLICATION_JSON)
 	public UserDTO getUser(@PathParam("userId") String userId) {
+		BackMeUpUser activeUser = ((BackmeupPrincipal)securityContext.getUserPrincipal()).getUser();
+		if(activeUser.getUserId() != Long.parseLong(userId)) {
+			throw new WebApplicationException(Status.FORBIDDEN);
+		}
+		
 		BackMeUpUser userModel = getLogic().getUserByUserId(userId);
 		return getMapper().map(userModel, UserDTO.class);
-//		return DummyDataManager.getUserDTO();
 	}
 	
+	@RolesAllowed("user")
 	@PUT
 	@Path("/{userId}")
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public UserDTO updateUser(@PathParam("userId") String userId, UserDTO user) {
+		BackMeUpUser activeUser = ((BackmeupPrincipal)securityContext.getUserPrincipal()).getUser();
+		if(activeUser.getUserId() != Long.parseLong(userId)) {
+			throw new WebApplicationException(Status.FORBIDDEN);
+		}
+		
 		BackMeUpUser userModel = getLogic().getUserByUserId(userId);
 		userModel.setFirstname(user.getFirstname());
 		userModel.setLastname(user.getLastname());
 		userModel.setEmail(user.getEmail());
+		
 		BackMeUpUser newUser = getLogic().updateUser(userModel);
 		return getMapper().map(newUser, UserDTO.class);
-//		return user;
 	}
 	
+	@RolesAllowed("user")
 	@DELETE
 	@Path("/{userId}")
 	@Produces(MediaType.APPLICATION_JSON)
 	public void deleteUser(@PathParam("userId") String userId) {
+		BackMeUpUser activeUser = ((BackmeupPrincipal)securityContext.getUserPrincipal()).getUser();
+		if(activeUser.getUserId() != Long.parseLong(userId)) {
+			throw new WebApplicationException(Status.FORBIDDEN);
+		}
+		
 		getLogic().deleteUser(userId);
 	}
 }
